requirements.txt
pyperclip==1.8.2
Pillow==10.1.0
pywin32==306; sys_platform == 'win32'  # Windows only
